Refactor: 프로필 뱃지 코드 구조 수정
